bump sensors datasize from u16b to 32b

+++ Src/sensors/include/iat.hpp
@@ -10,21 +10,21 @@ namespace IAT
      * -healt check
      * @return valores de 0 a 260
      */
-    uint16_t get_value(uint16_t);
+    int32_t get_value(int32_t);
 
     /** @brief devuelve valor mas exacto para calibracion inicial
      * dependiendo las opciones de compilacion se agrega:
      * -healt check
      * @return valores de -15 a 120
      */
-    uint32_t get_calibrate_value(uint16_t);
+    int32_t get_calibrate_value(int32_t);
 
     /** @brief revisa el estado del sensor para disparar un DTC en caso que sea necesario
      * @param in valor en Volt del sensor
      */
     uint8_t* dtc(void);
 
-    extern uint16_t last_value;
+    extern int32_t last_value;
 } // namespace IAT
 
 #endif
