+++ Src/sensors/src/map.cpp
@@ -6,14 +6,14 @@
 #include "../utils/basic_electronics.h"
 #include "../include/map.hpp"
 
-uint16_t MAP::get_value(uint16_t filt_input)
+int32_t MAP::get_value(int32_t filt_input)
 {
-    return (uint16_t)MAP_CAL(filt_input);
+    return (int32_t)MAP_CAL(filt_input);
 }
 
-uint32_t MAP::get_calibrate_value(uint16_t filt_input)
+int32_t MAP::get_calibrate_value(int32_t filt_input)
 {
-    return (uint32_t)MAP_CAL(filt_input) * 100;
+    return (int32_t)MAP_CAL(filt_input) * 100;
 }
 
 uint8_t *MAP::dtc()
