+++ Src/sensors/src/tps.cpp
@@ -1,23 +1,23 @@
 #include "../include/tps.hpp"
 #include "dtc_codes.h"
 
-uint16_t TPS::last_value = 0;
+int32_t TPS::last_value = 0;
 
-uint8_t TPS::get_value(uint16_t filt_input)
+int32_t TPS::get_value(int32_t filt_input)
 {
-    uint16_t _t = convert_to_volt(filt_input) * 1334;
+    int32_t _t = convert_to_volt(filt_input) * 1334;
     if (_t > TPS_MAX_A || _t < TPS_MIN_A)
         return 0;
-    TPS::last_value = (uint16_t)TPS_CALC_A(_t);
-    return (uint8_t)last_value;
+    TPS::last_value = (int32_t)TPS_CALC_A(_t);
+    return (int32_t)last_value;
 }
 
-uint32_t TPS::get_calibrate_value(uint16_t filt_input)
+int32_t TPS::get_calibrate_value(int32_t filt_input)
 {
-    uint16_t _t = convert_to_volt(filt_input) * 1334;
+    int32_t _t = convert_to_volt(filt_input) * 1334;
     if (_t > TPS_MAX_A || _t < TPS_MIN_A)
         return 0;
-    return (uint32_t)TPS_CALC_A(_t) * 1000;
+    return (int32_t)TPS_CALC_A(_t) * 1000;
 }
 
 uint8_t *TPS::dtc()
@@ -30,18 +30,18 @@ uint8_t *TPS::dtc()
     return 0;
 }
 
-uint8_t TPS::dtc(uint16_t in1, uint16_t in2)
+uint8_t TPS::dtc(int32_t in1, int32_t in2)
 {
     //TODO: laburar x1
     return 0;
 }
-uint8_t TPS::get_value(uint16_t filt_input, uint16_t filt_input2)
+int32_t TPS::get_value(int32_t filt_input, int32_t filt_input2)
 {
     //TODO: laburar x2
     return 0;
 }
 
-uint32_t TPS::get_calibrate_value(uint16_t filt_input, uint16_t filt_input2)
+int32_t TPS::get_calibrate_value(int32_t filt_input, int32_t filt_input2)
 {
     //TODO: laburar x3
     return 0;
