+++ Src/sensors/utils/basic_electronics.h
@@ -6,14 +6,14 @@
  *  @param in : valor del adc a convertir
  *  @return mV convertidos
  */
-uint16_t convert_to_volt(uint16_t);
+int32_t convert_to_volt(int32_t);
 
 /** @brief Convierte valor del ADC a resistencia:
  *  @param in : valor del adc a convertir
  *  @return resistencia en Ohm
  */
-uint16_t convert_to_resistance(uint16_t);
+int32_t convert_to_resistance(int32_t);
 
-uint16_t thermistor_get_temperature(int16_t);
+int32_t thermistor_get_temperature(int32_t);
 
 #endif
