+++ Src/sensors/utils/ema.h
@@ -8,14 +8,14 @@
 #define EMA_LOW_ALPHA 0.45 //< Valor de alpha, 0.32 a 0.50 recomendado para el filtro LP, variar dependiendo del PCB
 
 struct EMAHighPass{
-    uint16_t lastValue;   //< ultimo valor luego del filtrado, se utiliza este para acceder a la medicion
-    uint16_t actualValue; //< valor nuevo a filtrar
-    uint16_t lastLP;      //< este es el ultimo valor del filtro LowPass, luego se resta al valor actual para dejar la frecuencia alta
+    int32_t lastValue;   //< ultimo valor luego del filtrado, se utiliza este para acceder a la medicion
+    int32_t actualValue; //< valor nuevo a filtrar
+    int32_t lastLP;      //< este es el ultimo valor del filtro LowPass, luego se resta al valor actual para dejar la frecuencia alta
 };
 
 struct EMALowPass{
-    uint16_t lastValue;   //< ultimo valor luego del filtrado, se utiliza este para acceder a la medicion
-    uint16_t actualValue; //< valor nuevo a filtrar
+    int32_t lastValue;   //< ultimo valor luego del filtrado, se utiliza este para acceder a la medicion
+    int32_t actualValue; //< valor nuevo a filtrar
 };
 
 /** @brief  Exponential Moving Average, Low Pass Filter
